Moving KinOkHttpClientFactory & associated android specific test. Downgrading kin-base to java 6 compatibility.
