+++ kin-sdk/kin-sdk-lib/src/main/java/kin/sdk/KinOkHttpClientFactory.java
@@ -1,4 +1,4 @@
-package kin.base;
+package kin.sdk;
 
 import okhttp3.Interceptor;
 import okhttp3.OkHttpClient;
